# api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://127.0.0.1:8080
    description: Generated server url
paths:
  /api/resUsers/{id}:
    get:
      tags:
        - res-user-resource
      operationId: getResUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResUserDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
        - res-user-resource
      operationId: updateResUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResUserDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - res-user-resource
      operationId: deleteResUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /api/resLocations/{id}:
    get:
      tags:
        - res-location-resource
      operationId: getResLocation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResLocationDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
        - res-location-resource
      operationId: updateResLocation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResLocationDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - res-location-resource
      operationId: deleteResLocation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/resDistricts/{id}:
    get:
      tags:
        - res-district-resource
      operationId: getResDistrict
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResDistrictDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
        - res-district-resource
      operationId: updateResDistrict
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResDistrictDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - res-district-resource
      operationId: deleteResDistrict
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/resDistrictWards/{id}:
    get:
      tags:
        - res-district-ward-resource
      operationId: getResDistrictWard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResDistrictWardDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
        - res-district-ward-resource
      operationId: updateResDistrictWard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResDistrictWardDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - res-district-ward-resource
      operationId: deleteResDistrictWard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /api/resUsers:
    get:
      tags:
        - res-user-resource
      operationId: getAllResUsers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResUserDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      tags:
        - res-user-resource
      operationId: createResUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResUserDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /api/resLocations:
    get:
      tags:
        - res-location-resource
      operationId: getAllResLocations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResLocationDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      tags:
        - res-location-resource
      operationId: createResLocation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResLocationDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /api/resDistrictWards:
    get:
      tags:
        - res-district-ward-resource
      operationId: getAllResDistrictWards
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResDistrictWardDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      tags:
        - res-district-ward-resource
      operationId: createResDistrictWard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResDistrictWardDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
components:
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: "#/components/schemas/ApiFieldError"
    ApiFieldError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        property:
          type: string
        rejectedValue:
          type: object
        path:
          type: string
    ResUserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
          maxLength: 255
          minLength: 0
        password:
          type: string
          maxLength: 255
          minLength: 0
        name:
          type: string
          maxLength: 255
          minLength: 0
        role:
          type: array
          items:
            type: integer
            format: int64
      required:
        - username
...
    ResLocationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 255
          minLength: 0
        latitude:
          type: integer
          format: int64
        longitude:
          type: integer
          format: int64
        street:
          type: string
          maxLength: 255
          minLength: 0
        districWard:
          type: integer
          format: int64
        province:
          type: integer
          format: int64
        district:
          type: integer
          format: int64
    ResDistrictDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 255
          minLength: 0
        code:
          type: string
          maxLength: 255
          minLength: 0
        province:
          type: integer
          format: int64
      required:
        - code
        - name
        - province
    ResDistrictWardDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 255
          minLength: 0
        code:
          type: string
          maxLength: 255
          minLength: 0
        district:
          type: integer
          format: int64
      required:
        - code
        - district
        - name
...