# src/main/java/com/oop7hkem10/hrm/model/ResDistrictDTO.java
package com.oop7hkem10.hrm.model;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.Getter;
import lombok.Setter;


@Getter
@Setter
public class ResDistrictDTO {

    private Long id;

    @NotNull
    @Size(max = 255)
    private String name;

    @NotNull
    @Size(max = 255)
    @ResDistrictCodeUnique
    private String code;

    @NotNull
    private Long province;

}